+++ src/main/java/com/schoolhealth/schoolmedical/entity/MedicalEventMedicationUsed.java
@@ -16,7 +16,7 @@ public class MedicalEventMedicationUsed {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "medication_used_id")
-    private int medicationUsedId;
+    private Long medicationUsedId;
 
     @Column(name = "medication_id", nullable = false)
     private int medicationId;
