+++ src/main/java/com/schoolhealth/schoolmedical/repository/NotificationRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationRepo extends JpaRepository<UserNotification,Integer> {
+public interface NotificationRepo extends JpaRepository<UserNotification,Long> {
 }
